HIVE-25211: Create database throws NPE (#2362) (Yongzhi Chen, reviewed by Naveen Gangam)

+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/HMSHandler.java
@@ -1459,11 +1459,11 @@ public class HMSHandler extends FacebookBase implements IHMSHandler {
 
         if (db.getCatalogName() != null && !db.getCatalogName().
             equals(Warehouse.DEFAULT_CATALOG_NAME)) {
-          if (madeManagedDir) {
+          if (madeManagedDir && dbMgdPath != null) {
             wh.deleteDir(dbMgdPath, true, db);
           }
         } else {
-          if (madeManagedDir) {
+          if (madeManagedDir && dbMgdPath != null) {
             try {
               UserGroupInformation.getLoginUser().doAs(new PrivilegedExceptionAction<Void>() {
                 @Override public Void run() throws Exception {
@@ -1477,7 +1477,7 @@ public class HMSHandler extends FacebookBase implements IHMSHandler {
             }
           }
 
-          if (madeExternalDir) {
+          if (madeExternalDir && dbExtPath != null) {
             try {
               UserGroupInformation.getCurrentUser().doAs(new PrivilegedExceptionAction<Void>() {
                 @Override public Void run() throws Exception {
